Add restaurants, protect all endpoints by user and restaurant id, setup environment variables, setup categories as objects

+++ controller.js
@@ -1,23 +1,27 @@
 const db = require("./models");
 const Dish = db.Dish;
-const User= db.User;
+const User = db.User;
+const Restaurant = db.Restaurant;
 
 const jwt = require('jsonwebtoken');
 
 const passport = require('passport');
 const passportJWT = require('passport-jwt');
 
+const { JWT_SECRET } = require('./config.js')
+// import { JWT_SECRET } from './config.js';
+
 let ExtractJwt = passportJWT.ExtractJwt;
 
 let JwtStrategy = passportJWT.Strategy;
 let jwtOptions = {};
 jwtOptions.jwtFromRequest = ExtractJwt.fromAuthHeaderAsBearerToken();
-jwtOptions.secretOrKey = 'wowwow';
+jwtOptions.secretOrKey = JWT_SECRET;
 passReqToCallback: true;
 
 // auth strategy
 let strategy = new JwtStrategy(jwtOptions, function(jwt_payload, next) {
-  User.getUser({ email: jwt_payload.id })
+  User.getUser({ email: jwt_payload.email })
     .then(user => {
     if (user) {
       next(null, user);
@@ -29,13 +33,15 @@ let strategy = new JwtStrategy(jwtOptions, function(jwt_payload, next) {
 
 passport.use(strategy)
 
-
-
 // users
 const registerUser = (req, res) => {
   try {
-    let newUser = User.register(req.body.email, req.body.password);
-    res.send("User " + req.body.email + " was successfully created!");
+    let newUser;
+    User.register(req.body.email, req.body.password, req.body.phone, req.body.role)
+      .then((user) => {
+          res.send("User " + user.email + " was successfully created!");
+      });
+    
   } catch (err) {
     res.status(500).send(err);
   }
@@ -57,12 +63,48 @@ const loginUser = async (req, res) => {
 
 }
 
-
-const createDish = (req, res) => {
-  const dish = {
-    name: req.body.name
+// Restaurants
+const createRestaurant = (req, res) => {
+  const restaurant = {
+    name: req.body.name,
+    streetAddress: req.body.streetAddress,
+    city: req.body.city,
+    state: req.body.state,
+    zip: req.body.zip,
+    phone: req.body.phone,
+    url: req.body.url
   }
 
+  Restaurant.create(restaurant)
+    .then(data => {
+      res.send(data);
+    })
+    .catch(err => {
+      res.status(500).send({
+        message: err.message || "An error occured creating while processing this request"
+      });
+    });
+}
+
+// gets restaurant information based on authentication
+// const getRestaurant = (req, res) => {
+
+// }
+
+
+// Dishes
+// TODO: Get user from auth and get restaurant from user
+const createDish = (req, res) => {
+  const dish = {
+    name: req.body.name,
+    description: req.body.description,
+    addons: req.body.addons,
+    canRemove: req.body.canRemove,
+    notes: req.body.notes,
+    tableTalkPoints: req.body.tableTalkPoints,
+    restaurantId: req.user.restaurantId         // register to user's restaurant
+  }
+  
   Dish.create(dish)
     .then(data => {
       res.send(data);
@@ -75,7 +117,10 @@ const createDish = (req, res) => {
 };
 
 const dishesList = (req, res) => {
-  Dish.findAll()
+  userRestaurantId = req.user.restaurantId
+  Dish.findAll({
+    where: {restaurantId: userRestaurantId}
+  })
     .then(data => {
       res.send(data);
     })
@@ -89,9 +134,18 @@ const dishesList = (req, res) => {
 const getDish = (req, res) => {
   const id = req.params.id;
 
+  userRestaurantId = req.user.restaurantId
   Dish.findByPk(id)
-    .then(data => {
-      res.send(data);
+    .then(dish => {
+      // verify user belongs to restauraunt of dish requested
+      if(dish && dish.restaurantId == userRestaurantId) {
+        res.send(dish);
+      }
+      else {
+        res.status(404).send({
+          message: "Could not find dish"
+        });
+      }
     })
     .catch(err => {
       res.status(500).send({
@@ -101,15 +155,20 @@ const getDish = (req, res) => {
 };
 
 const updateDish = (req, res) => {
-  Dish.update(req.body,
-  {
-    where: {
-      id: req.params.id
+  userRestaurantId = req.user.restaurantId
+  Dish.findByPk(req.id)
+  .then(dish => {
+    // verify user belongs to restauraunt of dish to update
+    if(dish && dish.restaurantId == userRestaurantId) {
+      dish.updateDish(req.body);
+    }
+    else {
+      // sends if dish does not exist, or user does not have access
+      res.status(404).send({
+        message: "Could not find dish to update"
+      });
     }
   })
-  .then(res.send({
-    message: "Dish id=" + req.params.id + " updated successfully"
-  }))
   .catch(err => {
     res.status(500).send({
       message: err.message || "An error occured while updating dish with id=" + id
@@ -118,17 +177,36 @@ const updateDish = (req, res) => {
 };
 
 const deleteDish = (req, res) => {
-  Dish.destroy({
-    where: {id: id}
+  userRestaurantId = req.user.restaurantId
+  Dish.findByPk(req.id)
+  .then(dish => {
+    // verify user belongs to restauraunt of dish to update
+    if(dish && dish.restaurantId == userRestaurantId) {
+      Dish.destroy({
+        where: {id: id}
+      })
+      .then(res.send({
+        message: "Dish was deleted successfully"
+      }))
+      .catch(err => {
+        res.status(500).send({
+          message: err.message || "An error occured while deleting dish with id=" + id
+        });
+      });
+    }
+    else {
+      // sends if dish does not exist, or user does not have access
+      res.status(404).send({
+        message: "Could not find dish to update"
+      });
+    }
   })
-  .then(res.send({
-    message: "Dish was deleted successfully"
-  }))
   .catch(err => {
     res.status(500).send({
-      message: err.message || "An error occured while deleting     dish with id=" + id
+      message: err.message || "An error occured while updating dish with id=" + id
     });
   });
+
 };
 
 const getDishForMobile = (req, res) => {
@@ -144,5 +222,6 @@ module.exports = {
   getDishForMobile,
   registerUser,
   loginUser,
-  passport
+  passport,
+  createRestaurant
 }
